add prefix for all admin commands
With a prefix for all admin commands you don't have to worry about
accidentaly sending an admin command to the whole list if you make a
typo or something.

Now instead of sending "add <number>" you send "/add <number>". And
any messages starting with "/" that are not valid commands are swallowed.