remake to simpler group/keyword system
Now each group has a keyword and a phone it belongs to and there is
a send prefix (default #). So a sendout would look like "#kw message"
this would be delivered to the group with keyword "kw". Likewise for
admin commands, "/kw add 12345" would add the number 12345 to the group
with keyword "kw".

If a number is only in one group it can send messages without the send
prefix and keyword.